Simplify code, modularize it
Public methods in Alexandria lib now return a result, containing
either the result value or an error of the operation.

API paths in Alexandria binary now return an API response if the call
to the library fails.

API paths are now categorized by the first element of the path into
separated modules.

In both the binary and the library, new types are introduced in order
to simplify the type returned by functions.